refactor(biometric_data.tsx): added defaults.ts for default values and cleaned up files for PR
n/a
