+++ client/app/screen_one/page.tsx
@@ -12,7 +12,7 @@ import ConnectionStrength from "@/components/hmd_link/conn_strength";
 import { CameraFeed } from "@/components/hmd_link/camera_feed";
 import TodoLister from "@/components/hmd_link/todo-lister";
 import ContentManager from "@/components/general/content-manager";
-import EvaTelemetry from "@/components/hmd_link/eva_telemetry";
+import BiometricTelemetry from "@/components/hmd_link/biometric_data";
  
 const NoSSR_Timers = dynamic(() => import('@/components/ui/timing'), { ssr: false })
 
@@ -21,9 +21,9 @@ const HomePage = () => {
     <div className="h-full flex flex-row gap-x-8">
       <div className="flex flex-col items-left pl-3 justify-start max-w-[600px]">
         <NoSSR_Timers />
-        <EvaTelemetry evaNumber={1} bpm={0.0} temp={0.0} oxy={0.0} className="pt-[4.2rem]" />
+        <BiometricTelemetry evaNumber={1} bpm={0.0} temp={0.0} oxy={0.0} className="pt-[4.2rem]" />
         <CameraFeed />
-        <EvaTelemetry evaNumber={2} bpm={0.0} temp={0.0} oxy={0.0} />
+        <BiometricTelemetry evaNumber={2} bpm={0.0} temp={0.0} oxy={0.0} />
         <CameraFeed />
         <TodoLister />
       </div>
